+++ .storybook/main.js
@@ -3,6 +3,6 @@ module.exports = {
   addons: [
     "@storybook/addon-actions",
     "@storybook/addon-links",
-    "@storybook/addon-storysource"
-  ]
+    "@storybook/addon-storysource",
+  ],
 };