+++ src/components/Loading.js
@@ -14,7 +14,7 @@ const Loading = ({ children, className }) => {
 
 Loading.propTypes = {
   children: PropTypes.node,
-  className: PropTypes.string
+  className: PropTypes.string,
 };
 
 export default Loading;