+++ stories/components/ProjectsList.js
@@ -5,7 +5,7 @@ import { UI_URL } from "../settings";
 
 const ProjectsList = ({ data }) => (
   <ul>
-    {data.map(item => (
+    {data.map((item) => (
       <li key={item.slug}>
         <a href={`${UI_URL}/projects/#${item.slug}`}>{item.title}</a>
       </li>
@@ -17,9 +17,9 @@ ProjectsList.propTypes = {
   data: PropTypes.arrayOf(
     PropTypes.shape({
       title: PropTypes.string.isRequired,
-      slug: PropTypes.string.isRequired
+      slug: PropTypes.string.isRequired,
     })
-  )
+  ),
 };
 
 export default ProjectsList;